Put in a basic GraphicSystem skeleton.
This is a big commit. I put in the basic GraphicSystem flow and ran
into a few issues with that. It turns out my Sprite class (and all
the other Graphic components AND all other classes) needed to have
proper copy constructors, copy assignment, and swap operations. This
is because the ObjectPool relies on the copy assignment for
initialization of all its entries. Using the default copy constructor
resulted in the pointer members of the Sprite class to be shallow
copied, meaning that every instance in the pool pointed to the same
memory.

Secondly, I forgot to initialize the size of the graphics system
camera because I wrote the Scene class to automatically initialize
all cameras in the scene graph and I forgot that I need to do it
manually for the new camera.