Change BuilderSceneECS to have mouse actions modify map_root transform instead of graphics system camera.
It is much better to modify the map_root's scene graph transform instead of the camera because it
simplifies the code to create this scene. Instead of having to make two graphical systems, one for the
map root, and the other for the hud, you can leave the graphical system at its default. If I want to
add more layers at the map_root/hud_root level, I don't have to add more systems either.